Change likefollow to likefollowstar

+++ cosinnus/views/common.py
@@ -25,7 +25,7 @@ from cosinnus.models.tagged import LikeObject
 from cosinnus.utils.context_processors import cosinnus as cosinnus_context
 from cosinnus.utils.context_processors import settings as cosinnus_context_settings
 from cosinnus.utils.permissions import check_object_write_access, \
-    check_object_likefollow_access
+    check_object_likefollowstar_access
 from cosinnus.utils.urls import safe_redirect
 from cosinnus.views.mixins.group import RequireCreateObjectsInMixin
 
@@ -123,7 +123,7 @@ def cosinnus_logout(request, **kwargs):
 
 UNSPECIFIED = object()
 
-def apply_likefollow_object(obj, user, like=UNSPECIFIED, follow=UNSPECIFIED):
+def apply_likefollowstar_object(obj, user, like=UNSPECIFIED, follow=UNSPECIFIED, star=UNSPECIFIED):
     """
         Toggles the like or follow, or both states on a LikeObject.
         If no LikeObject existed, and either like or follow is True, create a new object.
@@ -138,10 +138,13 @@ def apply_likefollow_object(obj, user, like=UNSPECIFIED, follow=UNSPECIFIED):
     
     content_type = ContentType.objects.get_for_model(model_cls)
     liked_obj = get_object_or_None(LikeObject, content_type=content_type, object_id=obj.id, user=user)
-    if (not like or like is UNSPECIFIED) and (not follow or follow is UNSPECIFIED) and liked_obj is None:
+
+    if ((not like or like is UNSPECIFIED) and
+        (not follow or follow is UNSPECIFIED) and
+        (not star or star is UNSPECIFIED) and liked_obj is None):
         # only unlike or unfollow or both, and no object: nothing to do here
         pass
-    elif not (like is UNSPECIFIED and follow is UNSPECIFIED):
+    elif not (like is UNSPECIFIED and follow is UNSPECIFIED and star is UNSPECIFIED):
         if liked_obj is None:
             # initialize an object but don't save it yet
             liked_obj = LikeObject(content_type=content_type, object_id=obj.id, user=user, liked=False, followed=auto_follow)
@@ -150,47 +153,52 @@ def apply_likefollow_object(obj, user, like=UNSPECIFIED, follow=UNSPECIFIED):
             liked_obj.liked = like
         if not follow is UNSPECIFIED:
             liked_obj.followed = follow
+        if not star is UNSPECIFIED:
+            liked_obj.starred = star
         # check for deletion state
-        if not liked_obj.liked and (not liked_obj.followed or delete_if_unlike):
+        if not liked_obj.liked and (not liked_obj.followed or delete_if_unlike) and not liked_obj.starred:
             liked_obj.delete()
             liked_obj = None
         else:
             liked_obj.save()
-        
+
         # delete the objects like/folow cache
         obj.clear_likes_cache()
         # update the liked object's index
         if hasattr(obj, 'update_index'):
             obj.update_index()
-    
+
     was_liked = liked_obj and liked_obj.liked or False
     was_followed = liked_obj and liked_obj.followed or False
-    return was_liked, was_followed
+    was_stared = liked_obj and liked_obj.starred or False
+    return was_liked, was_followed, was_stared
 
 
 def apply_like_object(obj, user, like):
     # create, change or delete the LikeObj, but take care that the FOLLOW is false before deleting
-    return apply_likefollow_object(obj, user, like=like, follow=UNSPECIFIED)
+    return apply_likefollowstar_object(obj, user, like=like, follow=UNSPECIFIED)
 
 def apply_follow_object(obj, user, follow):
     # create, change or delete the LikeObj, but take care that the LIKE is false before deleting
-    return apply_likefollow_object(obj, user, like=UNSPECIFIED, follow=follow)
+    return apply_likefollowstar_object(obj, user, like=UNSPECIFIED, follow=follow)
 
 
 @csrf_protect
-def do_likefollow(request, **kwargs):
+def do_likefollowstar(request, **kwargs):
+
     """ Expected POST arguments:
         - ct: django content-type string (expects e.g. 'cosinnus_note.Note')
         - id: Id of the object. (optional if slug is given)
         - slug: Slug of the object (optional if id is given)
         - like: (optional) 0/1, whether to like or unlike
         - follow: (optional) 0/1, whether to follow or unfollow
+        - star: (optional) 0/1, whether to star or unstar
         
         User needs to be logged in.
         Target object needs to be visible (permissions) to the logged in user.
         If `follow`=1 param is given without `like`, a liked=False,followed=True object will be created.
         If the LikeObject results in being liked=False,followed=False, it will be deleted immediately.
-        Special for likefollow combined:
+        Special for likefollowstar combined:
             If the LikeObject results in being liked=False, no matter the follow state, it will be deleted immediately
     """
     
@@ -210,10 +218,11 @@ def do_likefollow(request, **kwargs):
     slug = request.POST.get('slug', None)
     like = PARAM_VALUE_MAP.get(request.POST.get('like', None), UNSPECIFIED)
     follow = PARAM_VALUE_MAP.get(request.POST.get('follow', None), UNSPECIFIED)
+    star = PARAM_VALUE_MAP.get(request.POST.get('star', None), UNSPECIFIED)
     
-    if ct is None or (id is None and slug is None) or (like is UNSPECIFIED and follow is UNSPECIFIED):
+    if ct is None or (id is None and slug is None) or (like is UNSPECIFIED and follow is UNSPECIFIED and star is UNSPECIFIED):
         return HttpResponseBadRequest('Incomplete data submitted.')
-    
+
     app_label, model = ct.split('.')
     model_cls = apps.get_model(app_label, model)
     
@@ -230,12 +239,12 @@ def do_likefollow(request, **kwargs):
     if obj is None:
         return HttpResponseNotFound('Target object not found on server.')
     
-    if not check_object_likefollow_access(obj, request.user):
+    if not check_object_likefollowstar_access(obj, request.user):
         return HttpResponseForbidden('Your access to this object is forbidden.')
     
-    was_liked, was_followed = apply_likefollow_object(obj, request.user, like=like, follow=follow)
+    was_liked, was_followed, was_starred = apply_likefollowstar_object(obj, request.user, like=like, follow=follow, star=star)
     
-    return JsonResponse({'liked': was_liked, 'followed': was_followed})
+    return JsonResponse({'liked': was_liked, 'followed': was_followed, 'starred': was_starred})
     
 
 class DeleteElementView(RequireCreateObjectsInMixin, View):
